+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ### Unreleased
 
+* Add a 2-dimensional, double-precision, floating-point vector named `Vector2d`.
+
 ### [v0.06](https://github.com/realityforge/vecmath/tree/v0.06) (2020-10-18)
 [Full Changelog](https://github.com/realityforge/vecmath/compare/v0.05...v0.06)